Add remote calculator project and according tests
This change demonstrates how we can integration-test a whole service (e.g. .NET web api). Keeping the given example of a calculator, a remote calculator was added. It implements the ICalculator interface but instead of calculating itself, it calls a POST route of the calculation service. During the lifetime of the RemoteCalculatorFixture, the .NET framework will spin up the service in the background and provide a configured HttpClient (see WebApplicationFactory.CreateClient method). We can use this client in the ConfigureServices method to pass it to our RemoteCalculator.